requirements.txt
streamlit>=1.32
google-generativeai>=0.6.0
